major refactor, create new 'core' subpackage, tests pass, refactor algorithms.cnmfe
