+++ mesmerize_napari/core/algorithms/cnmf.py
