+++ mesmerize_napari/core/caiman_extensions/__init__.py
@@ -1,0 +1,5 @@
+from .common import CaimanDataFrameExtensions, CaimanSeriesExtensions
+from .mcorr import MCorrExtensions
+from .cnmf import CNMFExtensions
+
+__all__ = ['CaimanDataFrameExtensions', 'CaimanSeriesExtensions', 'MCorrExtensions', 'CNMFExtensions']
