+++ mesmerize_napari/core/utils.py
